store: add architectureEndpoints for composerApi
