+++ src/store/api/backend/onprem/composerApi/helpers/index.ts
@@ -1,0 +1,1 @@
+export { getCloudConfigs } from './getCloudConfigs';
